backend/requirements.txt
flask==2.3.3
gunicorn==21.2.0
python-dotenv==1.0.0
redis==5.0.1
openai==1.30.0
httpx==0.27.0
marshmallow==3.20.1
flask-cors==4.0.0
pytest==7.4.3
black==23.10.1
isort==5.12.0
flask-jwt-extended==4.5.3
weaviate-client==4.5.4
google-auth==2.23.4
google-auth-oauthlib==1.1.0
google-auth-httplib2==0.1.1
Pillow==10.0.0
